+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Enable `ShippingSimulator` to accept and postal code format and validate using the country rules.
 
 ## [3.37.0] - 2019-05-17
 ### Added